parametersAndSessions

+++ Back_Net8_Pathfinder/Core/Entities/Asistencia.cs
@@ -14,25 +14,25 @@ public class Asistencia
     public DateTime AsisFecha { get; set; }
     [Required(ErrorMessage = "La frecuencia es obligatoria")]
     [Column(TypeName = "tinyint")]
-    public int AsisFrecuencia { get; set; }
+    public byte AsisFrecuencia { get; set; }
     [Required(ErrorMessage = "La devoción matutina es obligatoria")]
     [Column(TypeName = "tinyint")]
-    public int AsisDevocion { get; set; }
+    public byte AsisDevocion { get; set; }
     [Required(ErrorMessage = "La cuota es obligatoria")]
     [Column(TypeName = "tinyint")]
-    public int AsisCuota { get; set; }
+    public byte AsisCuota { get; set; }
     [Required(ErrorMessage = "La disciplina es obligatoria")]
     [Column(TypeName = "tinyint")]
-    public int AsisDisciplina { get; set; }
+    public byte AsisDisciplina { get; set; }
     [Required(ErrorMessage = "El año bíblico es obligatorio")]
     [Column(TypeName = "tinyint")]
-    public int AsisAnioBiblico { get; set; }
+    public byte AsisAnioBiblico { get; set; }
     [Required(ErrorMessage = "Los requerimientos son obligatorios")]
     [Column(TypeName = "tinyint")]
-    public int AsisRequerimiento { get; set; }
+    public byte AsisRequerimiento { get; set; }
     [Required]
     [Column(TypeName = "tinyint")]
-    public int AsisTotal { get; set; }
+    public byte AsisTotal { get; set; }
     [Required(ErrorMessage = "El monto es obligatorio")]
     [Column(TypeName = "decimal(9,3)")]
     public float AsisMonto { get; set; }
